More auto init shenanigans
Removed constructor tests. The initial idea was to run arbitrary things
before main. This was cool to run random steps before main but for a
plugin like system I wanted to test having some ordering (did not
actually sort Plugins based on priority but it is easy now if I need it) + making sure
we have thread allocation context initialized (still not threading but
wanted to kind of preapre this). Well, we can now do our main calls or
whatever and then call arbitrary deep plugin init functions. Not even
sure if I am going to use this but it was cool to test out.